refactor: update to import/export syntax
